# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Blog",
        "description": "Cette API permet de gérer un blog.",
        "contact": {
            "name": "Joséphin Sylvère",
            "email": "josephinsylvere@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Récupère tous les tags",
                "description": "Retourne une liste de tous les tags avec leurs posts associés.",
                "operationId": "getAllTags",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tag"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Crée un nouveau tag",
                "description": "Enregistre un nouveau tag avec les données fournies après validation.",
                "operationId": "createTag",
                "requestBody": {
                    "description": "Données du nouveau tag",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom du tag",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tag"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/tags/{id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Récupère un tag par son ID",
                "description": "Retourne un tag et ses posts associés en fonction de l'ID spécifié.",
                "operationId": "getTagById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du tag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tag"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Met à jour un tag existant",
                "description": "Met à jour un tag avec les données fournies après validation.",
                "operationId": "updateTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du tag à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données mises à jour pour le tag",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom du tag",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tag mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tag"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Supprime un tag",
                "description": "Supprime un tag spécifié par son ID et renvoie un message de succès.",
                "operationId": "deleteTag",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du tag à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le tag a été supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/tags/{id}/posts": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Récupère les posts associés à un tag spécifique",
                "description": "Retourne tous les posts associés à un tag spécifié par son ID.",
                "operationId": "getTagPosts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du tag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Tag": {
                "title": "Tag",
                "description": "Un modèle de tag pour le blog",
                "properties": {
                    "id": {
                        "description": "L'identifiant unique du tag",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Le nom du tag",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Le slug du tag",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "La date et l'heure de création du tag",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "La date et l'heure de la dernière mise à jour du tag",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "title": "Post",
                "description": "Un modèle de post pour le blog",
                "properties": {
                    "id": {
                        "description": "L'identifiant unique du post",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Le titre du post",
                        "type": "string"
                    },
                    "content": {
                        "description": "Le contenu du post",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "L'identifiant de l'utilisateur qui a créé le post",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "La date et l'heure de création du post",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "La date et l'heure de la dernière mise à jour du post",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "auth_bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez le jeton JWT de cette manière: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}